Small fix for build problem on linux/mac

+++ source/MaterialXShaderGen/ShaderGenerators/Glsl/OgsFx/OgsFxShaderGenerator.cpp
@@ -140,7 +140,7 @@ ShaderPtr OgsFxShaderGenerator::generate(const string& shaderName, ElementPtr el
     bool lighting = shader.hasClassification(SgNode::Classification::SHADER | SgNode::Classification::SURFACE);
 
     // Turn on fixed formatting since OgsFx doesn't support scientific values
-    Value::ScopedFloatFormatting fmt(std::ios_base::fixed);
+    Value::ScopedFloatFormatting fmt((int)std::ios_base::fixed);
 
     //
     // Emit code for vertex shader stage
